[!~] Packed source base. Sync..
First, [Ppm] is now part of [workshop]. So no implementation code
in this repo.

Second, [workshop] code is packed in one .lua file. This move have pros
and cons, but that's what I'm done. So demo/tools repos which
implementations were added to [workshop] should now move focus on their
code, not code of implementation.